Correctly finding units in MoreThanOneStrategy
Imagine the following scenario:

    TopUnitOne             TopUnitTwo
        |                     |
      SomeSchool             SomeOtherSchool
        |                      |
        |                     Deparment
        |                        |
        -------------------------
                  |
              DegreeUnit

And that we say to the more than one strategy that TopUnitOne,
SomeSchool and SomeOtherSchool are the 3 available units.

If we start in DegreeUnit the expected result is:

SomeSchool and SomeOtherSchool in case we allow multiple units
and
SomeSchool in case in don't

Notice that even when using multiple units TopUnitOne is not
expected since we found SomeSchool. This means that when finding a unit,
we do not continue the search to it's parents.

That constraint is that this patch is fixing, since before this the
result would be

TopUnitOne, SomeSchoo, SomeOtherSchool

which is incorrect.